Problem: catch_error's logic is convoluted and error-prone
In fact, I already had a bug caught there. Still, it doesn't make it any
more readable or easy to explain. It also relies on internal
implementation details of SubTransaction.

Solution: make SubTransaction roll back on drop instead of commit

This way, we can continue maintaining control over the sub-transaction
when everything was successful and commit when it is appropriate.

`catch_error` is gone completely and the code is seemingly much cleaner
now.