Problem: use of SpiClientWrapper type
Nobody would know what this type is and why is it there.

Solution: use Box<SpiClient> type instead

It doesn't solve the underlying problem (needing to have something that
is Deref/DerefMut to SpiClient; I still want to see if I can find a
better solution) it pbring a much more familiar interface to it.